Fix scanner build errors and remove hero scan-line
Scanner (preserves all detection logic — no modules removed):
- Cs2CheatScanModule clashed with CS2CheatScanModule on the
  Windows runner's case-insensitive filesystem, causing the
  uppercase module to vanish at checkout and CS0246 in ScanEngine.
  Rename the lite variant to Cs2BasicCheatScanModule (file + class)
  and update its one reference in ScanEngine.
- SeaOfThievesCheatScanModule:798 — missing $ prefix on the
  continuation string with {string.Join(", ", hits)} interpolation
  caused the literal comma to be parsed as an object-initializer
  member declarator (CS0747). Prefix all four continuation strings.
- EtwTamperScanModule:247 — ushort ord * 4 evaluates to int, then
  uint + int promoted to long for the RvaToOff(uint) argument
  (CS1503). Cast the offset to uint.

UI:
- Remove the cyan scan-line element from the landing hero per
  user request, and drop the now-unused keyframes + class.

+++ scanner/src/ZeroTrace.Core/Modules/EtwTamperScanModule.cs
@@ -244,7 +244,7 @@ public sealed class EtwTamperScanModule : IScanModule
                 if (ordOff < 0 || ordOff + 2 > diskBytes.Length) continue;
                 ushort ord = BitConverter.ToUInt16(diskBytes, ordOff);
 
-                int funcRvaOff = RvaToOff(diskBytes, secStart, numSec, addrOfFuncs + ord * 4);
+                int funcRvaOff = RvaToOff(diskBytes, secStart, numSec, addrOfFuncs + (uint)(ord * 4));
                 if (funcRvaOff < 0 || funcRvaOff + 4 > diskBytes.Length) continue;
                 uint funcRva = BitConverter.ToUInt32(diskBytes, funcRvaOff);
                 if (funcRva == 0) continue;
